Make window resizable and fix clipping bugs
Not sure why I thought the resize thing wasn't working properly before in minifb, but it actually works as expected, albeit with some interesting graphical glitches while resizing the window. This did uncover some additional clipping bugs in the rectangle drawing code, but those were easy to fix.

Interestingly, I'm also seeing a bug where if the mouse cursor is changed before hovering over the window client area, it keeps the changed cursor rather than using a pointer, which is a bit odd. This happens for example when you hover the mouse over the left edge of the window (where the cursor will show the resize left/right icon) and then onto the client area; the mouse cursor will keep showing as the resize left/right icon. It's probably just that no cursor was ever set for the client area or the window, so it just stays as-is. Will have to look into minifb for that one.